backend/requirements.txt
alembic==1.13.1
Babel==2.14.0
blinker==1.7.0
click==8.1.7
distlib==0.3.8
filelock==3.13.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
packaging==24.0
platformdirs==4.2.0
postgres==4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
psycopg2-pool==1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
setuptools==69.5.1
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
virtualenv==20.25.1
Werkzeug==3.0.2
WTForms==3.1.2
